55-57 + warnings

+++ p5460/print.cpp
@@ -5,7 +5,7 @@ void print_bool(bool b) {
 }
 
 void print_vec(std::vector<char> v) {
-    for (int i = 0; i < v.size(); i++) {
+    for (size_t i = 0; i < v.size(); i++) {
         std::cout<<v[i];
     }
 }
